+++ metadata.rb
@@ -1,9 +1,9 @@
-name             'hostname-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'hostname-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures hostname-nativex'
+description      'Installs/Configures hostname-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.2'
 
-depends		 "hostsfile"
+depends  'hostsfile'